pyproject.toml
[tool.poetry]
name = "questionnaire"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
django-nested-admin = "^4.1.1"
coverage = "^7.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
